requirements.txt
requests
fastapi
pydantic
asyncpg
uvicorn
numpy
Pillow
pandas
scikit_learn
matplotlib
